+++ cmd/login.go
@@ -5,99 +5,22 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"io"
 	"log"
 	"net/http"
-	"net/url"
 	"os"
 	"os/exec"
 	"runtime"
 	"time"
 
-	"github.com/adrg/xdg"
-	"github.com/alexkappa/mustache"
+	"github.com/amp-labs/cli/clerk"
 	"github.com/amp-labs/cli/vars"
-	"github.com/clerkinc/clerk-sdk-go/clerk"
 	"github.com/spf13/cobra"
 	"github.com/tidwall/pretty"
 )
 
-// HTML is shown to the user after they log in.
-var HTML = `<!doctype html>` + "\n" + //nolint:gochecknoglobals
-	`<html>
-        <head>
-                <meta charset="utf-8"/>
-                <title>Login</title>
-                <!--[if lt IE 9]>
-                <script src="//html5shim.googlecode.com/svn/trunk/html5.js"></script>
-                <![endif]-->
-        </head>
-        <body>
-            <h1>Successfully logged in as {{email}}</h1>
-        <h2>Please close this tab or page and return to the CLI</h2>
-        </body>
-</html>`
-
 const (
-	ServerPort                 = 3535
-	ClerkClientSessionPathDev  = "%s/v1/client?_clerk_js_version=4.50.1&__dev_session=%s"
-	ClerkClientSessionPathProd = "%s/v1/client?_clerk_js_version=4.50.1"
+	ServerPort = 3535
 )
-
-// loginData is the data that is stored in the jwt.json file.
-type loginData struct {
-	UserID    string            `json:"userId"`
-	SessionID string            `json:"sessionId"`
-	Token     string            `json:"token"`
-	Cookies   map[string]string `json:"cookies"`
-}
-
-type token struct {
-	Jwt string `json:"jwt"`
-}
-
-type verification struct {
-	Status string `json:"status"`
-}
-
-type email struct {
-	ID           string       `json:"id"`
-	Address      string       `json:"email_address"` //nolint:tagliatelle
-	Verification verification `json:"verification"`
-}
-
-type phone struct {
-	ID           string       `json:"id"`
-	Number       string       `json:"phone_number"` //nolint:tagliatelle
-	Verification verification `json:"verification"`
-}
-
-type user struct {
-	ID             string  `json:"id"`
-	Username       string  `json:"username"`
-	FirstName      string  `json:"first_name"`               //nolint:tagliatelle
-	LastName       string  `json:"last_name"`                //nolint:tagliatelle
-	ImageURL       string  `json:"image_url"`                //nolint:tagliatelle
-	PrimaryEmail   string  `json:"primary_email_address_id"` //nolint:tagliatelle
-	PrimaryPhone   string  `json:"primary_phone_number_id"`  //nolint:tagliatelle
-	EmailAddresses []email `json:"email_addresses"`          //nolint:tagliatelle
-	PhoneNumbers   []phone `json:"phone_numbers"`            //nolint:tagliatelle
-}
-
-type session struct {
-	LastActiveToken token `json:"last_active_token"` //nolint:tagliatelle
-	CreatedAt       int64 `json:"created_at"`        //nolint:tagliatelle
-	UpdatedAt       int64 `json:"updated_at"`        //nolint:tagliatelle
-	User            user  `json:"user"`
-}
-
-type response struct {
-	Sessions []session `json:"sessions"`
-}
-
-type clientResponse struct {
-	Response response `json:"response"`
-}
 
 type handler struct{}
 
@@ -142,138 +65,36 @@ const JwtFilePermissions = 0o600
 
 // processLogin takes the JWT token, verifies it, and then stores it in the jwt.json file.
 func processLogin(ctx context.Context, payload []byte, write bool) (string, string, error) { //nolint:cyclop
-	data := &loginData{}
+	data := &clerk.LoginData{}
 	if err := json.Unmarshal(payload, data); err != nil {
 		return "", "", err
 	}
 
-	path := getJwtPath()
+	path := clerk.GetJwtPath()
 	if write {
 		if err := os.WriteFile(path, pretty.Pretty(payload), JwtFilePermissions); err != nil {
 			return "", "", err
 		}
 	}
 
-	// Call out to clerk and ask for session info using the JWT token.
-	u := getClerkURL(data)
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, u, nil)
+	jwt, err := clerk.FetchJwt(ctx)
 	if err != nil {
 		return "", "", err
 	}
 
-	for k, v := range data.Cookies {
-		req.AddCookie(&http.Cookie{
-			Name:     k,
-			Value:    v,
-			Path:     "/",
-			Domain:   getClerkDomain(),
-			Secure:   true,
-			HttpOnly: true,
-		})
-	}
-
-	rsp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return "", "", err
-	}
-
-	defer func() { _ = rsp.Body.Close() }()
-
-	bb, err := io.ReadAll(rsp.Body)
-	if err != nil {
-		return "", "", err
-	}
-
-	if rsp.StatusCode != http.StatusOK {
-		return "", "", fmt.Errorf("http %d (%s)", rsp.StatusCode, string(bb)) //nolint:goerr113
-	}
-
-	cr := &clientResponse{}
-	if err := json.Unmarshal(bb, cr); err != nil {
-		return "", "", err
-	}
-
-	if len(cr.Response.Sessions) == 0 {
-		return "", "", fmt.Errorf("no sessions found in response") //nolint:goerr113
-	}
-
-	return decodeJWT(cr.Response.Sessions[0].LastActiveToken.Jwt)
-}
-
-func getClerkURL(data *loginData) string {
-	if vars.Stage == "prod" {
-		return fmt.Sprintf(ClerkClientSessionPathProd, vars.ClerkRootURL)
-	}
-
-	return fmt.Sprintf(ClerkClientSessionPathDev, vars.ClerkRootURL, data.Token)
-}
-
-func getClerkDomain() string {
-	u, err := url.Parse(vars.ClerkRootURL)
-	if err != nil {
-		log.Fatalln(err)
-	}
-
-	return u.Hostname()
-}
-
-func decodeJWT(jwt string) (string, string, error) {
-	// Using a dummy value here because DecodeToken doesn't actually use the secret.
-	c, err := clerk.NewClient("dummy")
-	if err != nil {
-		return "", "", err
-	}
-
-	// Extract the claims (which includes the email address) from the JWT token.
-	claims, err := c.DecodeToken(jwt)
-	if err != nil {
-		return "", "", err
-	}
-
-	// Grab the email address from the claims.
-	emailStr, ok := claims.Extra["email"].(string)
-	if !ok {
-		return "", "", fmt.Errorf("couldn't find email address in claims") //nolint:goerr113
-	}
-
-	ht, err := getHTML(emailStr)
-	if err != nil {
-		return "", "", err
-	}
-
-	// Return the HTML and email
-	return ht, emailStr, nil
-}
-
-func getHTML(emailStr string) (string, error) {
-	// Render the HTML
-	tmpl := mustache.New()
-	if err := tmpl.ParseString(HTML); err != nil {
-		return "", err
-	}
-
-	ht, err := tmpl.RenderString(map[string]string{
-		"email": emailStr,
-	})
-	if err != nil {
-		return "", err
-	}
-
-	return ht, nil
+	return clerk.DecodeJWT(jwt)
 }
 
 const ReadHeaderTimeoutSeconds = 3
 
 // loginCmd represents the login command.
 var loginCmd = &cobra.Command{ //nolint:gochecknoglobals
-	Use:    "login",
-	Short:  "Log in to an Ampersand account",
-	Long:   "Log in to an Ampersand account.",
-	Hidden: true,
+	Use:   "login",
+	Short: "Log in to an Ampersand account",
+	Long:  "Log in to an Ampersand account.",
 	Run: func(cmd *cobra.Command, args []string) {
 		needLogin := false
-		path := getJwtPath()
+		path := clerk.GetJwtPath()
 		fileInfo, err := os.Stat(path)
 		if err != nil {
 			if os.IsNotExist(err) {
@@ -317,24 +138,6 @@ var loginCmd = &cobra.Command{ //nolint:gochecknoglobals
 			os.Exit(0)
 		}
 	},
-}
-
-func getJwtName() string {
-	if vars.Stage == "prod" {
-		return "amp/jwt.json"
-	}
-
-	return fmt.Sprintf("amp/jwt-%s.json", vars.Stage)
-}
-
-// getJwtPath returns the path to the jwt.json file where the JWT token is stored.
-func getJwtPath() string {
-	path, err := xdg.ConfigFile(getJwtName())
-	if err != nil {
-		log.Fatalln(err)
-	}
-
-	return path
 }
 
 // openBrowser tries to open the URL in a browser. Should work on most standard platforms.
